`refactor`: maior conformidade com arquitetura MVC
- Add pacotes para `controllers`, `views` e `forms`.
- Separa lógica de autenticação em `auth_controller` e `views.auth`.
- Separa cada classe de formulário em arquivos diferentes para melhor
  manutenabilidade.
- Agrupa todas as extensões usadas na aplicação em `extensions.py` e
  reestrutura dependências nos pacotes que as usam a fim de evitar
  imports circulares.
- Altera a rota de `auth/register` para `/register`, o que faz mais
  sentido para a aplicação.

Essa refatoração melhora a modularidade e a organização do código.
