+++ src/app/__init__.py
@@ -3,16 +3,10 @@ import logging
 from logging.handlers import RotatingFileHandler
 
 from flask import Flask
-from flask_bootstrap import Bootstrap5
-from flask_wtf import CSRFProtect
 
 from .config import configurations
-from .models import db, migrate
-from .auth import auth, bcrypt, login_manager
-from .views import views
-
-bootstrap = Bootstrap5()
-csrf = CSRFProtect()
+from .extensions import db, init_extensions
+from .views import views, auth
 
 
 def create_app(config='default') -> Flask:
@@ -20,18 +14,13 @@ def create_app(config='default') -> Flask:
     app.config.from_object(configurations[config])
 
     # Inicialização de extensões
-    csrf.init_app(app)
-    db.init_app(app)
-    migrate.init_app(app, db)
-    login_manager.init_app(app)
-    bcrypt.init_app(app)
-    bootstrap.init_app(app)
+    init_extensions(app=app)
 
     # Registro de Blueprints
     app.register_blueprint(views, url_prefix='/')
     app.register_blueprint(auth, url_prefix='/auth')
 
-    # Inicia o logger em prod
+    # Inicia o logger em prod/testing
     if not app.debug:
         configure_logging(app)
