+++ src/app/controllers/__init__.py
@@ -1,0 +1,7 @@
+from .auth_controller import login_user_controller
+from .register_controller import register_user_controller
+
+__all__ = [
+    'login_user_controller',
+    'register_user_controller'
+]
