+++ src/app/forms/__init__.py
@@ -1,0 +1,7 @@
+from .login_form import LoginForm
+from .registration_form import RegistrationForm
+
+__all__ = [
+    'LoginForm',
+    'RegistrationForm'
+]
