+++ src/app/forms/login_form.py
@@ -1,0 +1,11 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, PasswordField, SubmitField, BooleanField
+from wtforms.validators import DataRequired, Email
+
+
+class LoginForm(FlaskForm):
+    email = StringField(
+        "Email", validators=[DataRequired(), Email()])
+    senha = PasswordField("Senha", validators=[DataRequired()])
+    lembre_de_mim = BooleanField("Mantenha-me conectado")
+    submit = SubmitField("Login")
