+++ src/app/models/comentario.py
@@ -1,6 +1,6 @@
 from datetime import datetime, timezone
 
-from .setup import db
+from ..extensions import db
 
 
 class Comentario(db.Model):
