+++ src/app/models/medico.py
@@ -1,4 +1,4 @@
-from .setup import db
+from ..extensions import db
 
 
 class Medico(db.Model):
