--- src/app/models/setup.py
@@ -1,5 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-from flask_migrate import Migrate
-
-db = SQLAlchemy()
-migrate = Migrate()
