+++ src/app/templates/base.html
@@ -60,7 +60,7 @@
               <a class="nav-link" href="{{ url_for('auth.login') }}">Login</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="{{ url_for('auth.register') }}"
+              <a class="nav-link" href="{{ url_for('views.register') }}"
                 >Registrar</a
               >
             </li>
