+++ src/app/templates/register.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %} {% from 'bootstrap5/form.html' import render_form %}
 {% block content %}
 <div class="container">
-  <form method="POST" action="{{ url_for('auth.register') }}">
+  <form method="POST" action="{{ url_for('views.register') }}">
     {{ render_form(form) }}
   </form>
 </div>
