--- src/app/views.py
@@ -1,16 +1,0 @@
-from flask import Blueprint, render_template
-from flask_login import login_required
-
-views = Blueprint('views', __name__)
-
-
-@views.route('/')
-@views.route('/home')
-def home():
-    return render_template('home.html')
-
-
-@views.route('/account')
-@login_required
-def account():
-    return render_template('account.html', title='Conta')
