+++ src/app/views/__init__.py
@@ -1,0 +1,7 @@
+from .views import views
+from .auth import auth
+
+__all__ = [
+    'views',
+    'auth'
+]
